requirements.txt
# Core ML stack — versions pinned to the environment used for all experiments.
# Do not bump any pinned version mid-experiment; update only after all four arms are complete.
torch==2.8.0
transformers==5.5.3
accelerate==1.13.0
trl==1.0.0
peft==0.18.1
bitsandbytes==0.49.2   # 4-bit quantization for smoke runs on 8 GB GPU
vllm==0.11.0           # fast inference for pass@1024 eval

# Pinned for byte-stable Hub downloads + Arrow/cache behavior alongside configs/math_beyond_hf_revision.txt
datasets==4.8.4
huggingface_hub==1.10.1

# sympy.parsing.latex requires antlr4 4.11.x specifically
antlr4-python3-runtime==4.11.1
sympy

# Numerics and data handling
numpy<2.0.0
pandas
scipy
pyyaml
tqdm

# Evaluation and analysis
scikit-learn
umap-learn
matplotlib
seaborn

# Logging / experiment tracking
tensorboard
wandb
